Bivarite MGL and survival MGL copula

Density, distribution function, and random generation for the bivariate MGL and survival MGL copula with copula parameter \(\delta\).

dcMGL.bivar(u1, u2, pars)

pcMGL.bivar(u1, u2, pars)

rcMGL.bivar(n, pars)

dcMGL180.bivar(u1, u2, pars)

pcMGL180.bivar(u1, u2, pars)

rcMGL180.bivar(n, pars)

Arguments

u1, u2
numeric vectors of equal length with values in \([0,1]\).

pars
copula parameter, denoted by \(\delta >0\).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dcMGL.bivar, pcMGL.bivar and rcMGL.bivar gives values of Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter \(\delta>0\).

dcMGL180.bivar, pcMGL180.bivar and rcMGL180.bivar gives values of Density, distribution function, and random generation for the d-dimensional MGL copula with copula parameter \(\delta>0\).

Details

For \(d=2\), the MGL copula density function is given by

\( c^{\text{MGL}}(u_1, \dots, u_d;a)=\frac{\Gamma(a)^{d-1}\Gamma(a+\frac{d}{2})}{\Gamma(a+\frac{1}{2})^d}\frac{\prod_{j=1}^{d}(t(u_j;a)+1)^{a+\frac{1}{2}}}{\left(\sum_{j=1}^{d}t(u_j;a)+1\right)^{a+\frac{d}{2}}}, \quad (u_1, ..., u_d)\in \left[0,1\right]^d, \)

where \(a = \frac{1}{\delta}\), and \(I_{m,n}()\) denotes the beta cumulative distribution function (or regularized incomplete beta function) with parameters shape1 = m and shape2 = n implemented by R's pbeta.

The MGL copula distribution function has the form of

$$ C^{\text{MGL}}(u_1, \dots, u_d;a) =\mathbb{E}_{\Theta}\left[\prod_{j=1}^{d}\text{erfc}\left( \sqrt{t(u_j;a)\Theta} \right)\right] $$ with \(\Theta \sim \text{gamma}(a,1)\) and $$ t(u_j;a)=\frac{I^{-1}_{\frac{1}{2},{a}}(1-u_j)}{1-I^{-1}_{\frac{1}{2},{a}}(1-u_j)}. $$

For \(d=2\), the density and distribution function of survival MGL copula is $$ \bar{c}^{MGL}(u_1,u_2;\delta)=c^{MGL}(1-u_1,1-u_2;\delta), $$ and $$ \bar{C}^{MGL}(u_1,u_2;\delta)=u_1+u_2-1+C^{MGL}(1-u_1,1-u_2;\delta). $$

Examples

dcMGL.bivar(u1 = 0.001, u2 = 0.999, pars = 1)

#> [1] 0.05692685

pcMGL.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)

#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!

#> [1] 0.28685669 0.09779747 0.66292758

head(rcMGL.bivar(n = 200, pars = 0.8))

#>             [,1]        [,2]
#> [1,] 0.080750138 0.175027481
#> [2,] 0.834333037 0.956196217
#> [3,] 0.600760886 0.283132275
#> [4,] 0.157208442 0.026944806
#> [5,] 0.007399441 0.007099827
#> [6,] 0.466393497 0.863828114

dcMGL180.bivar(u1 = 0.001, u2 = 0.999, pars = 1)

#> [1] 0.05692685

pcMGL180.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)

#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!

#> [1] 0.2791833 0.0928438 0.6965791

rcMGL180.bivar(n = 200, pars = 0.8)

#>               [,1]       [,2]
#>   [1,] 0.666176683 0.63818027
#>   [2,] 0.250727348 0.29563001
#>   [3,] 0.086841989 0.77289720
#>   [4,] 0.795670937 0.37627969
#>   [5,] 0.528244389 0.13119396
#>   [6,] 0.614816900 0.15905236
#>   [7,] 0.607248071 0.11639091
#>   [8,] 0.781661570 0.79972353
#>   [9,] 0.992791059 0.91190428
#>  [10,] 0.636287857 0.10241464
#>  [11,] 0.473893270 0.60264141
#>  [12,] 0.155836607 0.19948551
#>  [13,] 0.864008829 0.36330395
#>  [14,] 0.623532669 0.14644196
#>  [15,] 0.736691352 0.02687033
#>  [16,] 0.502147030 0.84090262
#>  [17,] 0.980474749 0.08690073
#>  [18,] 0.912816190 0.52763745
#>  [19,] 0.004037336 0.88503537
#>  [20,] 0.057867795 0.44549699
#>  [21,] 0.875626549 0.81658667
#>  [22,] 0.884810698 0.53203455
#>  [23,] 0.633311896 0.85516916
#>  [24,] 0.528959964 0.34576472
#>  [25,] 0.215827554 0.24282854
#>  [26,] 0.509607187 0.35859995
#>  [27,] 0.898049045 0.98020768
#>  [28,] 0.056540276 0.27788635
#>  [29,] 0.293860498 0.29788596
#>  [30,] 0.497735763 0.50569710
#>  [31,] 0.470065224 0.44014038
#>  [32,] 0.741573652 0.35536621
#>  [33,] 0.288246583 0.10132919
#>  [34,] 0.031537450 0.22256146
#>  [35,] 0.338922210 0.53106977
#>  [36,] 0.203749410 0.48968522
#>  [37,] 0.926790868 0.94333143
#>  [38,] 0.115257116 0.46673019
#>  [39,] 0.173299623 0.35581117
#>  [40,] 0.274236041 0.87156219
#>  [41,] 0.632670978 0.25689030
#>  [42,] 0.476571583 0.41126437
#>  [43,] 0.530121087 0.34117424
#>  [44,] 0.901433614 0.93272438
#>  [45,] 0.483965090 0.06356204
#>  [46,] 0.550266173 0.32917989
#>  [47,] 0.450538452 0.42601005
#>  [48,] 0.350493810 0.70147142
#>  [49,] 0.840965174 0.02210793
#>  [50,] 0.706249208 0.88299200
#>  [51,] 0.108715133 0.86614090
#>  [52,] 0.411921374 0.19677220
#>  [53,] 0.366768437 0.29886657
#>  [54,] 0.740490524 0.76204305
#>  [55,] 0.681787273 0.08533981
#>  [56,] 0.660746119 0.83491158
#>  [57,] 0.239238861 0.30525280
#>  [58,] 0.989127536 0.98447371
#>  [59,] 0.936862661 0.44403502
#>  [60,] 0.736158208 0.16857254
#>  [61,] 0.406163411 0.12909780
#>  [62,] 0.702439971 0.20901208
#>  [63,] 0.905377228 0.98627401
#>  [64,] 0.238744640 0.17265014
#>  [65,] 0.227523152 0.09531818
#>  [66,] 0.598009607 0.38777151
#>  [67,] 0.940917047 0.80075735
#>  [68,] 0.249752116 0.06723795
#>  [69,] 0.471677645 0.34906141
#>  [70,] 0.724194882 0.59122119
#>  [71,] 0.738762450 0.11422073
#>  [72,] 0.456168831 0.20804779
#>  [73,] 0.800999221 0.89475432
#>  [74,] 0.354945228 0.40458436
#>  [75,] 0.249760154 0.79701686
#>  [76,] 0.729022166 0.51995315
#>  [77,] 0.124378091 0.38455416
#>  [78,] 0.148256826 0.66826934
#>  [79,] 0.954150689 0.39043434
#>  [80,] 0.618943141 0.05155573
#>  [81,] 0.816687946 0.67313999
#>  [82,] 0.078200359 0.40888339
#>  [83,] 0.093813151 0.01523369
#>  [84,] 0.807056682 0.71456430
#>  [85,] 0.135624723 0.46320401
#>  [86,] 0.986727398 0.97885918
#>  [87,] 0.114339660 0.85004675
#>  [88,] 0.608457893 0.61109915
#>  [89,] 0.923823687 0.94932304
#>  [90,] 0.353218695 0.22598814
#>  [91,] 0.083346103 0.03640293
#>  [92,] 0.165491380 0.27794299
#>  [93,] 0.498752930 0.49069158
#>  [94,] 0.397855230 0.40997785
#>  [95,] 0.531040966 0.62738422
#>  [96,] 0.113625762 0.31065308
#>  [97,] 0.329451508 0.48006429
#>  [98,] 0.606222281 0.10974291
#>  [99,] 0.448709328 0.19371890
#> [100,] 0.757507324 0.87219008
#> [101,] 0.796791171 0.77123544
#> [102,] 0.931658297 0.48587803
#> [103,] 0.692726263 0.57109464
#> [104,] 0.006881906 0.26757191
#> [105,] 0.883690823 0.42105210
#> [106,] 0.284576923 0.87483231
#> [107,] 0.773891269 0.04849870
#> [108,] 0.866651452 0.84536507
#> [109,] 0.072718238 0.06078325
#> [110,] 0.105877343 0.42019104
#> [111,] 0.796335608 0.07804549
#> [112,] 0.743017081 0.59755087
#> [113,] 0.385994396 0.58253572
#> [114,] 0.558500812 0.49573825
#> [115,] 0.684405315 0.03547557
#> [116,] 0.898686239 0.83660769
#> [117,] 0.727353804 0.72812576
#> [118,] 0.346300603 0.12128808
#> [119,] 0.072129898 0.71682485
#> [120,] 0.973375330 0.45378635
#> [121,] 0.440508803 0.57661871
#> [122,] 0.145826348 0.65699814
#> [123,] 0.924116950 0.24484767
#> [124,] 0.794171144 0.89091534
#> [125,] 0.504110353 0.19158759
#> [126,] 0.738143056 0.67342766
#> [127,] 0.584038549 0.03441286
#> [128,] 0.560850000 0.84394588
#> [129,] 0.817622081 0.62566122
#> [130,] 0.921671734 0.68308020
#> [131,] 0.982230044 0.45238548
#> [132,] 0.271820027 0.92540920
#> [133,] 0.029154975 0.90467135
#> [134,] 0.464782412 0.18789753
#> [135,] 0.833509232 0.85426388
#> [136,] 0.941824349 0.97783688
#> [137,] 0.540860981 0.12729469
#> [138,] 0.377201179 0.86968546
#> [139,] 0.405180799 0.50947495
#> [140,] 0.256981763 0.02227061
#> [141,] 0.475719866 0.39107944
#> [142,] 0.503174750 0.22605644
#> [143,] 0.985281197 0.97329681
#> [144,] 0.042080875 0.31074232
#> [145,] 0.546910765 0.79010713
#> [146,] 0.018468151 0.47059097
#> [147,] 0.455294834 0.77841879
#> [148,] 0.639957599 0.85934992
#> [149,] 0.496853895 0.58517119
#> [150,] 0.925366368 0.09302620
#> [151,] 0.030208110 0.10264842
#> [152,] 0.351498686 0.66619021
#> [153,] 0.015434533 0.28488756
#> [154,] 0.463866743 0.82592894
#> [155,] 0.238940595 0.59716408
#> [156,] 0.435614273 0.33778488
#> [157,] 0.221545696 0.47578780
#> [158,] 0.360939011 0.25425626
#> [159,] 0.621039264 0.10897619
#> [160,] 0.640469329 0.87324753
#> [161,] 0.155768584 0.31042088
#> [162,] 0.543250819 0.14898297
#> [163,] 0.820965709 0.71538752
#> [164,] 0.033790804 0.31429802
#> [165,] 0.286588048 0.46835291
#> [166,] 0.975113969 0.97446357
#> [167,] 0.373543506 0.42910267
#> [168,] 0.172078678 0.77613276
#> [169,] 0.356930968 0.30510330
#> [170,] 0.291515738 0.77007228
#> [171,] 0.129272847 0.41600673
#> [172,] 0.718187418 0.90669196
#> [173,] 0.826775158 0.58852219
#> [174,] 0.890329981 0.97682320
#> [175,] 0.413777808 0.15570938
#> [176,] 0.149278673 0.62376612
#> [177,] 0.078473824 0.66076059
#> [178,] 0.324338098 0.53001452
#> [179,] 0.649411838 0.36993225
#> [180,] 0.991341925 0.20406956
#> [181,] 0.569101207 0.62786037
#> [182,] 0.275064903 0.31899342
#> [183,] 0.587743613 0.16926467
#> [184,] 0.449137605 0.79847171
#> [185,] 0.843639430 0.04366499
#> [186,] 0.956998707 0.97906121
#> [187,] 0.315355887 0.68191364
#> [188,] 0.485597503 0.33008575
#> [189,] 0.054312152 0.41849568
#> [190,] 0.156903058 0.65973024
#> [191,] 0.935116907 0.03857171
#> [192,] 0.237889085 0.29141645
#> [193,] 0.921615748 0.96305363
#> [194,] 0.866492337 0.48767260
#> [195,] 0.190303799 0.10291951
#> [196,] 0.440616761 0.06229158
#> [197,] 0.266270314 0.50442167
#> [198,] 0.650304524 0.50367328
#> [199,] 0.017331333 0.55577152
#> [200,] 0.591146120 0.72112409