The GLMGA distribution

Density (dGLMGA), distribution function (pGLMGA), quantile function (qGLMGA) and random generation (rGLMGA) for the GLMGA distribution with parameters sigma, a and b.

dGLMGA(y, sigma, a, b, log = FALSE)

pGLMGA(y, sigma, a, b)

qGLMGA(u, sigma, a, b)

rGLMGA(n, sigma, a, b)

Arguments

y

vector of quantiles.

sigma

parameter of GLMGA distribution.

a

parameter of GLMGA distribution.

b

parameter of GLMGA distribution.

log

logical; if TRUE, probabilities/densities p are returned as log(p).

u

vector of probabilities.

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dGLMGA gives the density, pGLMGA gives the distribution function, qGLMGA gives the quantile function, and rGLMGA generates random deviates. Invalid arguments will result in return value NaN, with a warning. The length of the result is determined by n for rgamma, and is the maximum of the lengths of the numerical arguments for the other functions. The numerical arguments other than n are recycled to the length of the result. Only the first elements of the logical arguments are used.

Details

The GLMGA distribution with parameters (sigma, a, b) has density $$f(y)=\frac{(2b)^a}{\sigma B(a,\frac{1}{2})}\frac{y^{-(\frac{1}{2\sigma}+1)}}{(y^{-\frac{1}{\sigma}}+2b)^{a + \frac{1}{2}}},$$ for \(y>0,\sigma>0, a>0, b>0\).

The cumulative distribution function \(F(y)\) is

$$F(y)=1-I_{\frac{1}{2},a}(\frac{y^{-1/\sigma}}{y^{-1/\sigma}+2b}).$$

Here \(I_{m,n}()\) is the beta cumulative distribution function (or regularized incomplete beta function) with parameters shape1 = m and shape2 = n implemented by R's pbeta and defined in its help.

The quantile function \(F^{-1}(u)\) is

$$(2b)^{-\sigma}[\frac{I^{-1}_{\frac{1}{2},a}(1-p)}{1-I^{-1}_{\frac{1}{2},a}(1-p)}]^{-\sigma},$$ where \(u \in (0,1)\), and \(I_{m,n}^{-1}()\) denotes the inverse of the beta cumulative distribution function (or regularized incomplete beta function) with parameters shape1 = m and shape2 = n implemented by R's qbeta.

References

Zhengxiao Li, Jan Beirlant, Shengwang Meng. Generalizing The Log-Moyal Distribution And Regression Models For Heavy-Tailed Loss Data. ASTIN Bulletin: The Journal of the IAA, 11(1):57-99, 2021.

Examples

# density function at value 0.5 and 0.1
dGLMGA(c(0.5, 0.1), sigma = 2, a = 2, b = 3, log = FALSE)
#> [1] 0.2145154 0.9447679
# cdf at value 10 and 20.
pGLMGA(c(10, 20), sigma = 2, a = 2, b = 3)
#> [1] 0.6699700 0.7190817
# quantile function at level 50% and 10%
qGLMGA(c(0.5, 0.1), sigma = 2, a = 2, b = 3)
#> [1] 1.47657307 0.02151758
# simulate 10 samples from GLMGA distribution with parameters (2, 2, 3)
rGLMGA(n = 10, sigma = 2, a = 2, b = 3)
#>  [1] 13.290132993  1.415915018  1.689678049 50.206359928  2.023805530
#>  [6]  0.598293087  0.004544456  1.000779114 34.113737901  0.627872658