Bivarite MGB2 copula

Density, distribution function, and h-functions for the bivariate MGB2 copula proposed in Yang et al.,(2011).

dcMGB2.bivar(u1, u2, pars1, pars2, pars3)

pcMGB2.bivar(u1, u2, pars1, pars2, pars3)

hcMGB2.bivar(u1, u2, pars1, pars2, pars3)

Arguments

u1, u2
numeric vectors of equal length with values in \([0,1]\).

pars1, pars2, pars3
copula parameters, denoted by \(p_1, p_2, q\).

Value

dcMGB2.bivar gives the density

pcMGB2.bivar gives the distribution function

hcMGB2.bivar gives the h-functions (hfunc1, hfunc2).

Details

The MGB2 copula with parameters (p1, p2, q) has joint density $$c(u_1,u_2;p_1,p_2,q)=\frac{\Gamma(q)\Gamma(\sum_{i=1}^2 p_i + q)}{\prod_{i=1}^2\Gamma(p_i+q)}\frac{\prod_{i=1}^2 (1 + x(u_i;p_i,q))^{p_i+q}}{(1 + \sum_{i=1}^{2}x(u_i;p_i,q))^{\sum_{i=1}^{2}p_i+q}},$$ in which \(x(u_i;p_i,q) = I_{p_{i},q}^{-1}(u_i)/\left(1 - I_{p_{i},q}^{-1}(u_i)\right)\) for \(p_1, p_2>0, q>0\). (Here Gamma(a) is the function implemented by R's gamma and defined in its help.

The joint cdf of the MGB2 copula is

$$C(u_1,u_2;p_1,p_2,q)=\int_{0}^{+\infty}\prod_{i=1}^{2}G_p(\frac{I_{p_i,q}^{-1}(u_i)}{(1-I_{p_i,q}^{-1}(u_i))\theta})\times \frac{\theta^{-(q+1)}e^{-1/\theta}}{\Gamma(q)}d\theta, $$ where \(I_{m,n}^{-1}()\) denotes the inverse of the beta cumulative distribution function (or regularized incomplete beta function) with parameters shape1 = m and shape2 = n implemented by R's qbeta.

The h-function is defined as the conditional distribution function of a bivariate copula, i.e., $$h_1(u_2|u_1,p_1,p_2,q) := P(U_2 \leq u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,$$

$$h_2(u_1|u_2,p_1,p_2,q) := P(U_1 \leq u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,$$

where \((U_1, U_2) \sim C\), and \(C\) is a bivariate copula distribution function with parameter(s) \(p_1,p_2,q\).

References

Xipei Yang, Edward W Frees, and Zhengjun Zhang. A generalized beta copula with applications in modeling multivariate long-tailed data. Insurance: Mathematics and Economics, 49(2):265-284, 2011.

Examples

dcMGB2.bivar(u1 = c(0.2, 0.5), u2 = c(0.9, 0.2), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)

#> [1] 0.7182072 1.0750778

pcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)

#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!

#> [1] 0.47122605 0.01175683

hcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)$hfunc1

#> [1] 0.9322133 0.1173278

hcMGB2.bivar(u1 = c(0.5, 0.1), u2 = c(0.9, 0.1), pars1 = 0.5, pars2 = 0.5, pars3 = 1.5)$hfunc2

#> [1] 0.3717317 0.1173278