Conditional Distribution Function of a Bivariate MGL and survival Copula

Conditional Distribution Function of a Bivariate MGL and survival Copula

hcMGL.bivar(u1, u2, pars)

hcMGL180.bivar(u1, u2, pars)

Arguments

u1
numeric vectors of equal length with values in \([0,1]\).

u2
numeric vectors of equal length with values in \([0,1]\).

pars
numeric; single number or vector of size length(u1); copula parameter > 0.

Value

hcMGL.bivar returns a list with

hfunc1: \(\partial C(u_1,u_2) / \partial u_1,\)

hfunc2: \(\partial C(u_1,u_2) / \partial u_2,\)

Details

The h-function is defined as the conditional distribution function of a bivariate copula, i.e., $$h_1(u_2|u_1,\delta) := P(U_2 \leq u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,$$

$$h_2(u_1|u_2,\delta) := P(U_1 \leq u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,$$

where \((U_1, U_2) \sim C\), and \(C\) is a bivariate copula distribution function with parameter(s) \(\delta\). For more details see Aas et al. (2009).

The h-function corresponding to the bivariate MGL copula is given by $$ \begin{align} h_{2|1}^{\text{MGL}}(u_2|u_1;\delta)&= 1-I_{\frac{1}{2},\frac{1}{\delta}+\frac{1}{2}}\left[ \frac{t(u_2;\delta)}{t(u_1;\delta) + t(u_2;\delta)+1} \right],\\ h_{1|2}^{\text{MGL}}(u_1|u_2;\delta) &= 1-I_{\frac{1}{2},\frac{1}{\delta}+\frac{1}{2}}\left[ \frac{t(u_1;\delta)}{t(u_1;\delta) + t(u_2;\delta)+1} \right], \end{align} $$ for all \((u_1, u_2)\in\left[0,1\right]^2\). Here $$ t(u_j;a)=\frac{I^{-1}_{\frac{1}{2},{a}}(1-u_j)}{1-I^{-1}_{\frac{1}{2},{a}}(1-u_j)}. $$

The h-function corresponding to the survival bivariate MGL copula is given by $$ \begin{align} h_{2|1}^{\text{MGL180}}(u_2|u_1;\delta)&=1-h_{2|1}^{\text{MGL}}(1-u_2|1-u_1;\delta),\\ h_{1|2}^{\text{MGL180}}(u_1|u_2;\delta)&=1-h_{1|2}^{\text{MGL}}(1-u_1|1-u_2;\delta). \end{align} $$

References

Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009). Pair-copula constructions of multiple dependence. Insurance: Mathematics and Economics 44 (2), 182-198.

Examples

hcMGL.bivar(u1 = c(0.1, 0.001, 0.3), u2 = c(0, 0.9999, 0.88), pars = 2)

#> $hfunc1
#> [1] 0.0000000 0.9999998 0.9137196
#> 
#> $hfunc2
#> [1] 0.9999900000 0.0000012337 0.1123161318
#> 

hcMGL180.bivar(u1 = c(0.1, 0.001, 0.3), u2 = c(0, 0.9999, 0.88), pars = 2)

#> $hfunc1
#> [1] 0.0000000 1.0000000 0.9778406
#> 
#> $hfunc2
#> [1] 1.564345e-01 2.467403e-07 9.504334e-02
#>