Bivarite MGL and survival MGL copula

Density, distribution function, and random generation for the bivariate MGL/survival MGL copula with copula parameter \(\delta\).

dcMGL.bivar(u1, u2, pars)

pcMGL.bivar(u1, u2, pars)

rcMGL.bivar(n, pars)

dcMGL180.bivar(u1, u2, pars)

pcMGL180.bivar(u1, u2, pars)

rcMGL180.bivar(n, pars)

Arguments

u1, u2

numeric vectors of equal length with values in \([0,1]\).

pars

copula parameter, denoted by \(\delta >0\).

n

number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dcMGL.bivar, pcMGL.bivar and rcMGL.bivar gives values of density, distribution function, and random generation for the two dimensional MGL copula with copula parameter \(\delta>0\).

dcMGL180.bivar, pcMGL180.bivar and rcMGL180.bivar gives values of density, distribution function, and random generation for the two dimensional survival MGL copula with copula parameter \(\delta>0\).

Examples

# density function
dcMGL.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
#> [1] 0.05692685
# distribution function
pcMGL.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> [1] 0.28685669 0.09779747 0.66292758
# random generation
rcMGL.bivar(n = 200, pars = 0.8)
#>               [,1]        [,2]
#>   [1,] 0.498979587 0.106318620
#>   [2,] 0.244519977 0.616754049
#>   [3,] 0.849872394 0.886474510
#>   [4,] 0.959902704 0.315496407
#>   [5,] 0.848224422 0.488344156
#>   [6,] 0.161853314 0.447838237
#>   [7,] 0.814226017 0.929029844
#>   [8,] 0.106560829 0.399019727
#>   [9,] 0.838841253 0.355894824
#>  [10,] 0.391810386 0.640268838
#>  [11,] 0.324837977 0.678096389
#>  [12,] 0.957494021 0.591399860
#>  [13,] 0.525662444 0.431899290
#>  [14,] 0.645938196 0.997323971
#>  [15,] 0.476060096 0.609206519
#>  [16,] 0.186416988 0.316671322
#>  [17,] 0.249194316 0.171853075
#>  [18,] 0.360865783 0.198512268
#>  [19,] 0.763156285 0.812022075
#>  [20,] 0.883755429 0.199032800
#>  [21,] 0.119026884 0.362849864
#>  [22,] 0.836095966 0.256729653
#>  [23,] 0.277507921 0.725954771
#>  [24,] 0.939754824 0.224657673
#>  [25,] 0.641614174 0.864003231
#>  [26,] 0.164209957 0.318832532
#>  [27,] 0.257304299 0.673409387
#>  [28,] 0.531031482 0.321435645
#>  [29,] 0.196476672 0.668811915
#>  [30,] 0.596846078 0.802759147
#>  [31,] 0.310342207 0.961174598
#>  [32,] 0.846389622 0.093932079
#>  [33,] 0.934319621 0.434623768
#>  [34,] 0.103587764 0.364485495
#>  [35,] 0.855142909 0.873899793
#>  [36,] 0.388687518 0.820754533
#>  [37,] 0.748930260 0.678609759
#>  [38,] 0.838140539 0.578412612
#>  [39,] 0.565960393 0.295847474
#>  [40,] 0.810247482 0.784511655
#>  [41,] 0.207533240 0.562850365
#>  [42,] 0.387401803 0.535912397
#>  [43,] 0.278540215 0.159025174
#>  [44,] 0.097060916 0.614843987
#>  [45,] 0.770511752 0.931722204
#>  [46,] 0.816780510 0.070250696
#>  [47,] 0.651722493 0.410161295
#>  [48,] 0.493070401 0.937537482
#>  [49,] 0.424502076 0.035943650
#>  [50,] 0.504823408 0.411841273
#>  [51,] 0.934250870 0.093400177
#>  [52,] 0.519127023 0.099286312
#>  [53,] 0.806060045 0.606540917
#>  [54,] 0.056322502 0.076677066
#>  [55,] 0.282096692 0.441879259
#>  [56,] 0.819555340 0.908512253
#>  [57,] 0.652434058 0.446108167
#>  [58,] 0.527387074 0.112400135
#>  [59,] 0.413314314 0.487011354
#>  [60,] 0.681375117 0.799060632
#>  [61,] 0.073713708 0.196233212
#>  [62,] 0.739429664 0.734266920
#>  [63,] 0.732576140 0.435161324
#>  [64,] 0.630933655 0.018685767
#>  [65,] 0.662358838 0.217933458
#>  [66,] 0.918686993 0.584294521
#>  [67,] 0.311432532 0.666013811
#>  [68,] 0.657890151 0.537112392
#>  [69,] 0.961991525 0.384213336
#>  [70,] 0.369690342 0.729174655
#>  [71,] 0.068699073 0.554474723
#>  [72,] 0.699035052 0.529761222
#>  [73,] 0.846262637 0.795691134
#>  [74,] 0.635925480 0.993523848
#>  [75,] 0.827519583 0.692260928
#>  [76,] 0.212843298 0.368067867
#>  [77,] 0.316389729 0.774149997
#>  [78,] 0.153362677 0.223788090
#>  [79,] 0.289782682 0.856053216
#>  [80,] 0.086440613 0.016206511
#>  [81,] 0.049389686 0.299604986
#>  [82,] 0.675643123 0.372279407
#>  [83,] 0.419925393 0.945170755
#>  [84,] 0.865746873 0.438665582
#>  [85,] 0.134978237 0.944353249
#>  [86,] 0.267571348 0.880758147
#>  [87,] 0.292870587 0.554692113
#>  [88,] 0.536004341 0.513194023
#>  [89,] 0.839160871 0.779192063
#>  [90,] 0.448156950 0.691339080
#>  [91,] 0.576532981 0.401854665
#>  [92,] 0.963599941 0.036946736
#>  [93,] 0.302993362 0.241391998
#>  [94,] 0.156651257 0.194427584
#>  [95,] 0.550066399 0.779636672
#>  [96,] 0.876897175 0.426826983
#>  [97,] 0.895177718 0.204458238
#>  [98,] 0.601073844 0.672577184
#>  [99,] 0.869954692 0.576480578
#> [100,] 0.670540733 0.437279202
#> [101,] 0.345105531 0.498870574
#> [102,] 0.984997070 0.360788079
#> [103,] 0.494799427 0.288571893
#> [104,] 0.022792695 0.211865386
#> [105,] 0.160455651 0.253255863
#> [106,] 0.368422549 0.896453920
#> [107,] 0.005619131 0.004802758
#> [108,] 0.577659790 0.818091237
#> [109,] 0.708753643 0.882655169
#> [110,] 0.412133923 0.107451289
#> [111,] 0.228516463 0.160978302
#> [112,] 0.179839368 0.347258498
#> [113,] 0.492998485 0.579455560
#> [114,] 0.575672223 0.163308894
#> [115,] 0.627992291 0.271632954
#> [116,] 0.368151163 0.743597216
#> [117,] 0.629960129 0.344052257
#> [118,] 0.995157500 0.731356494
#> [119,] 0.797163696 0.680942704
#> [120,] 0.664879422 0.488361494
#> [121,] 0.842579084 0.816847329
#> [122,] 0.029340901 0.429811489
#> [123,] 0.487195255 0.636004217
#> [124,] 0.810414499 0.643101251
#> [125,] 0.354172420 0.363218359
#> [126,] 0.315458812 0.918434364
#> [127,] 0.358682382 0.434006463
#> [128,] 0.118651225 0.558723929
#> [129,] 0.750316634 0.971958260
#> [130,] 0.182803355 0.230212415
#> [131,] 0.775693493 0.726416512
#> [132,] 0.981168822 0.624555959
#> [133,] 0.405309833 0.632263146
#> [134,] 0.725281236 0.904423186
#> [135,] 0.355806959 0.166809262
#> [136,] 0.097780033 0.039553515
#> [137,] 0.370007030 0.255828148
#> [138,] 0.308236508 0.359119905
#> [139,] 0.113447592 0.593189251
#> [140,] 0.357049776 0.577538354
#> [141,] 0.873630065 0.918356763
#> [142,] 0.542427496 0.812507074
#> [143,] 0.871813429 0.562335456
#> [144,] 0.700518921 0.139125942
#> [145,] 0.103222956 0.443868664
#> [146,] 0.920086554 0.621192734
#> [147,] 0.300923429 0.558935178
#> [148,] 0.739876225 0.045431832
#> [149,] 0.264742284 0.530953807
#> [150,] 0.833943373 0.504263622
#> [151,] 0.054962836 0.127582111
#> [152,] 0.184686590 0.768948799
#> [153,] 0.207825083 0.585579082
#> [154,] 0.608945427 0.075601769
#> [155,] 0.370580385 0.652201212
#> [156,] 0.502490813 0.708253321
#> [157,] 0.424182494 0.884664447
#> [158,] 0.301545456 0.759601027
#> [159,] 0.345213843 0.157630737
#> [160,] 0.419877256 0.650551615
#> [161,] 0.327543092 0.847254133
#> [162,] 0.722996710 0.357003059
#> [163,] 0.664495864 0.438617169
#> [164,] 0.338008513 0.025502941
#> [165,] 0.961441203 0.236261494
#> [166,] 0.199633088 0.290421472
#> [167,] 0.385004645 0.300996527
#> [168,] 0.179690130 0.071792783
#> [169,] 0.176080706 0.667109381
#> [170,] 0.553973922 0.690209117
#> [171,] 0.985322519 0.443499546
#> [172,] 0.985796325 0.300856317
#> [173,] 0.928515952 0.181844298
#> [174,] 0.450476471 0.985444002
#> [175,] 0.416459794 0.093733992
#> [176,] 0.874301260 0.277801341
#> [177,] 0.454569529 0.403851354
#> [178,] 0.196438914 0.112787516
#> [179,] 0.428520609 0.646211578
#> [180,] 0.411672334 0.108772400
#> [181,] 0.907965088 0.764448504
#> [182,] 0.915983605 0.883740638
#> [183,] 0.472197015 0.756253163
#> [184,] 0.947241690 0.534660694
#> [185,] 0.144181256 0.527325537
#> [186,] 0.818012170 0.301492926
#> [187,] 0.952736185 0.539264971
#> [188,] 0.815753937 0.518427851
#> [189,] 0.373795208 0.104601241
#> [190,] 0.611659304 0.721916746
#> [191,] 0.204684082 0.705148678
#> [192,] 0.629274258 0.436158089
#> [193,] 0.925592247 0.658282437
#> [194,] 0.908056495 0.379434795
#> [195,] 0.063348734 0.729264113
#> [196,] 0.002795536 0.005833444
#> [197,] 0.201554239 0.952976778
#> [198,] 0.490326114 0.915185529
#> [199,] 0.487393226 0.456495592
#> [200,] 0.692427914 0.495014635
dcMGL180.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
#> [1] 0.05692685
pcMGL180.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> [1] 0.2791833 0.0928438 0.6965791
rcMGL180.bivar(n = 200, pars = 0.8)
#>              [,1]       [,2]
#>   [1,] 0.52202787 0.29339636
#>   [2,] 0.30042819 0.27705295
#>   [3,] 0.43176137 0.98344768
#>   [4,] 0.23483280 0.46675222
#>   [5,] 0.43354095 0.08360057
#>   [6,] 0.60034714 0.58858824
#>   [7,] 0.97796774 0.97911189
#>   [8,] 0.73811328 0.50375221
#>   [9,] 0.74983861 0.13681901
#>  [10,] 0.28387621 0.30997164
#>  [11,] 0.19711284 0.04845647
#>  [12,] 0.31589844 0.53245284
#>  [13,] 0.30312176 0.02243814
#>  [14,] 0.15110370 0.78691240
#>  [15,] 0.55647382 0.05059309
#>  [16,] 0.56944162 0.29778546
#>  [17,] 0.92903791 0.92531471
#>  [18,] 0.53326132 0.84340749
#>  [19,] 0.67797737 0.92802836
#>  [20,] 0.65308610 0.74045609
#>  [21,] 0.20335274 0.93143938
#>  [22,] 0.43962606 0.58901938
#>  [23,] 0.65470546 0.46825684
#>  [24,] 0.07261060 0.17371851
#>  [25,] 0.84419251 0.74720358
#>  [26,] 0.31633811 0.11878205
#>  [27,] 0.16281901 0.53108864
#>  [28,] 0.17829210 0.07039334
#>  [29,] 0.53188631 0.05306472
#>  [30,] 0.62268210 0.86694627
#>  [31,] 0.59545978 0.27264515
#>  [32,] 0.39773014 0.25729112
#>  [33,] 0.16634948 0.18881615
#>  [34,] 0.98836258 0.79339598
#>  [35,] 0.04152470 0.04662295
#>  [36,] 0.99742098 0.98996793
#>  [37,] 0.15831860 0.91568076
#>  [38,] 0.13001195 0.38573328
#>  [39,] 0.50613306 0.47683981
#>  [40,] 0.30190553 0.74500837
#>  [41,] 0.20023138 0.74389284
#>  [42,] 0.95010902 0.62178890
#>  [43,] 0.94389626 0.65633194
#>  [44,] 0.35921807 0.03065821
#>  [45,] 0.81645015 0.08510795
#>  [46,] 0.31347575 0.35433419
#>  [47,] 0.17091856 0.03643425
#>  [48,] 0.75575847 0.31453959
#>  [49,] 0.19385772 0.21267468
#>  [50,] 0.05568947 0.35861783
#>  [51,] 0.70373839 0.25643270
#>  [52,] 0.25633168 0.08234359
#>  [53,] 0.18535253 0.58582514
#>  [54,] 0.17969892 0.54255481
#>  [55,] 0.25544268 0.31383976
#>  [56,] 0.65987025 0.46341639
#>  [57,] 0.35464120 0.46513295
#>  [58,] 0.48673077 0.56709318
#>  [59,] 0.61750594 0.71174307
#>  [60,] 0.54232801 0.43628621
#>  [61,] 0.55250604 0.89261037
#>  [62,] 0.02147277 0.37766558
#>  [63,] 0.25648213 0.19866031
#>  [64,] 0.68889807 0.26810453
#>  [65,] 0.32933903 0.60046554
#>  [66,] 0.77800367 0.28899600
#>  [67,] 0.31077083 0.73132604
#>  [68,] 0.42213521 0.49298337
#>  [69,] 0.52679812 0.22779451
#>  [70,] 0.08839084 0.20128877
#>  [71,] 0.14650607 0.34888952
#>  [72,] 0.29657486 0.50431484
#>  [73,] 0.46870478 0.20870611
#>  [74,] 0.46399998 0.68022669
#>  [75,] 0.66713094 0.26429395
#>  [76,] 0.01336388 0.37058688
#>  [77,] 0.13921242 0.93728167
#>  [78,] 0.30024486 0.37793602
#>  [79,] 0.05115601 0.02049270
#>  [80,] 0.77789704 0.60395247
#>  [81,] 0.95757354 0.95245215
#>  [82,] 0.39626768 0.71107281
#>  [83,] 0.24409825 0.09178997
#>  [84,] 0.45484230 0.89245996
#>  [85,] 0.48574583 0.14216244
#>  [86,] 0.83971059 0.11899151
#>  [87,] 0.78575644 0.75396078
#>  [88,] 0.45292223 0.56393019
#>  [89,] 0.28135664 0.82308953
#>  [90,] 0.48280705 0.47205745
#>  [91,] 0.66820204 0.48427869
#>  [92,] 0.23319579 0.24068556
#>  [93,] 0.80484651 0.83971070
#>  [94,] 0.91846002 0.39807814
#>  [95,] 0.71090513 0.61403129
#>  [96,] 0.75663023 0.48359619
#>  [97,] 0.15526859 0.14183057
#>  [98,] 0.65942851 0.76677035
#>  [99,] 0.22425427 0.59616349
#> [100,] 0.10610894 0.39113946
#> [101,] 0.83887996 0.62756819
#> [102,] 0.45167521 0.01171873
#> [103,] 0.90213921 0.68854697
#> [104,] 0.35483445 0.68987472
#> [105,] 0.39577759 0.88213454
#> [106,] 0.13250950 0.04754528
#> [107,] 0.21226787 0.41803135
#> [108,] 0.57233828 0.18950354
#> [109,] 0.70248586 0.74741357
#> [110,] 0.05865328 0.40463560
#> [111,] 0.60645139 0.10591587
#> [112,] 0.25287408 0.36108152
#> [113,] 0.69351376 0.60676280
#> [114,] 0.37174054 0.51879159
#> [115,] 0.85887378 0.15331108
#> [116,] 0.65732601 0.74432933
#> [117,] 0.43093679 0.22340174
#> [118,] 0.09764108 0.20855938
#> [119,] 0.80055925 0.86827562
#> [120,] 0.15116002 0.47563504
#> [121,] 0.83520853 0.70053369
#> [122,] 0.86397592 0.04715144
#> [123,] 0.79605392 0.74332831
#> [124,] 0.70754927 0.21783605
#> [125,] 0.68134149 0.96267218
#> [126,] 0.56002550 0.84925539
#> [127,] 0.72548462 0.67968596
#> [128,] 0.22655841 0.06942341
#> [129,] 0.98295690 0.52892197
#> [130,] 0.84089094 0.03756287
#> [131,] 0.56039679 0.75325590
#> [132,] 0.50547667 0.19969644
#> [133,] 0.67951943 0.38770432
#> [134,] 0.80992093 0.26994636
#> [135,] 0.04054683 0.86021753
#> [136,] 0.42389055 0.27564411
#> [137,] 0.59998536 0.44786064
#> [138,] 0.31606926 0.78181805
#> [139,] 0.80028881 0.88405536
#> [140,] 0.08953614 0.57246077
#> [141,] 0.33418217 0.72979057
#> [142,] 0.37689859 0.43618418
#> [143,] 0.08929609 0.56912376
#> [144,] 0.18884857 0.61415923
#> [145,] 0.47002290 0.16913492
#> [146,] 0.32285138 0.01264893
#> [147,] 0.45323720 0.14899715
#> [148,] 0.69800714 0.69932241
#> [149,] 0.16833203 0.13135473
#> [150,] 0.82893665 0.11312099
#> [151,] 0.07409581 0.09463118
#> [152,] 0.12526673 0.15109888
#> [153,] 0.40368954 0.81098672
#> [154,] 0.15528714 0.07935658
#> [155,] 0.42059205 0.10960784
#> [156,] 0.87743710 0.90530715
#> [157,] 0.59835004 0.44872527
#> [158,] 0.70143197 0.94071184
#> [159,] 0.98645109 0.50786929
#> [160,] 0.13445135 0.79665524
#> [161,] 0.15779338 0.67009316
#> [162,] 0.14967183 0.59109215
#> [163,] 0.13848122 0.27599032
#> [164,] 0.63477016 0.24290422
#> [165,] 0.82206371 0.24905449
#> [166,] 0.32815316 0.80311203
#> [167,] 0.31410773 0.31235096
#> [168,] 0.22149379 0.24502525
#> [169,] 0.62239716 0.29358124
#> [170,] 0.30487280 0.12032447
#> [171,] 0.80795876 0.02537136
#> [172,] 0.79189449 0.92245193
#> [173,] 0.34115221 0.09208579
#> [174,] 0.17663452 0.86073879
#> [175,] 0.31450059 0.59583294
#> [176,] 0.65848460 0.66284129
#> [177,] 0.46165732 0.52431398
#> [178,] 0.12318600 0.56806862
#> [179,] 0.44079623 0.22464012
#> [180,] 0.51150715 0.39799132
#> [181,] 0.50310928 0.71853971
#> [182,] 0.83923062 0.75119417
#> [183,] 0.03179630 0.23809223
#> [184,] 0.83134364 0.95337372
#> [185,] 0.36037521 0.45407013
#> [186,] 0.44792497 0.94780908
#> [187,] 0.10551552 0.39276943
#> [188,] 0.58578346 0.01010250
#> [189,] 0.78596911 0.80595086
#> [190,] 0.04840975 0.46226343
#> [191,] 0.06935894 0.37079435
#> [192,] 0.67681983 0.10875182
#> [193,] 0.58028078 0.90762819
#> [194,] 0.48942214 0.76755017
#> [195,] 0.82779762 0.25963102
#> [196,] 0.85223607 0.23369731
#> [197,] 0.89046445 0.86287610
#> [198,] 0.50587873 0.29482925
#> [199,] 0.57157079 0.30389830
#> [200,] 0.63335828 0.55014704