Bivarite MGL and survival MGL copula

Density, distribution function, and random generation for the bivariate MGL/survival MGL copula with copula parameter \(\delta\).

dcMGL.bivar(u1, u2, pars)

pcMGL.bivar(u1, u2, pars)

rcMGL.bivar(n, pars)

dcMGL180.bivar(u1, u2, pars)

pcMGL180.bivar(u1, u2, pars)

rcMGL180.bivar(n, pars)

Arguments

u1, u2
numeric vectors of equal length with values in \([0,1]\).

pars
copula parameter, denoted by \(\delta >0\).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Value

dcMGL.bivar, pcMGL.bivar and rcMGL.bivar gives values of density, distribution function, and random generation for the two dimensional MGL copula with copula parameter \(\delta>0\).

dcMGL180.bivar, pcMGL180.bivar and rcMGL180.bivar gives values of density, distribution function, and random generation for the two dimensional survival MGL copula with copula parameter \(\delta>0\).

Examples

# density function
dcMGL.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
#> [1] 0.05692685
# distribution function
pcMGL.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> [1] 0.28685669 0.09779747 0.66292758
# random generation
rcMGL.bivar(n = 200, pars = 0.8)
#>               [,1]        [,2]
#>   [1,] 0.080750138 0.175027481
#>   [2,] 0.834333037 0.956196217
#>   [3,] 0.600760886 0.283132275
#>   [4,] 0.157208442 0.026944806
#>   [5,] 0.007399441 0.007099827
#>   [6,] 0.466393497 0.863828114
#>   [7,] 0.497777389 0.263612215
#>   [8,] 0.289767245 0.206527687
#>   [9,] 0.732881987 0.084325603
#>  [10,] 0.772521511 0.953634922
#>  [11,] 0.874600661 0.335919042
#>  [12,] 0.174940627 0.724829625
#>  [13,] 0.034241333 0.071396254
#>  [14,] 0.320385731 0.870131774
#>  [15,] 0.402328238 0.322122990
#>  [16,] 0.195669835 0.979761813
#>  [17,] 0.403538117 0.602814231
#>  [18,] 0.063661457 0.580665369
#>  [19,] 0.388701313 0.838855267
#>  [20,] 0.975547835 0.803908344
#>  [21,] 0.289892295 0.397548905
#>  [22,] 0.678380427 0.207946554
#>  [23,] 0.735319599 0.142766840
#>  [24,] 0.195956733 0.591148362
#>  [25,] 0.980539675 0.667836478
#>  [26,] 0.741521529 0.707821400
#>  [27,] 0.051446276 0.118734935
#>  [28,] 0.530212464 0.981426239
#>  [29,] 0.695823879 0.755595589
#>  [30,] 0.688556003 0.890960936
#>  [31,] 0.031230325 0.959250029
#>  [32,] 0.225562535 0.187771054
#>  [33,] 0.300830806 0.647498172
#>  [34,] 0.636465615 0.457454855
#>  [35,] 0.479024550 0.078237981
#>  [36,] 0.432171258 0.635075892
#>  [37,] 0.706433838 0.654338729
#>  [38,] 0.948576576 0.494438268
#>  [39,] 0.180338768 0.822086593
#>  [40,] 0.216899876 0.457470467
#>  [41,] 0.680162918 0.982250894
#>  [42,] 0.498845611 0.038050158
#>  [43,] 0.641679348 0.715240897
#>  [44,] 0.660284349 0.443133007
#>  [45,] 0.096024158 0.870424903
#>  [46,] 0.765600164 0.727218541
#>  [47,] 0.769674804 0.716366292
#>  [48,] 0.990712312 0.798283303
#>  [49,] 0.970520903 0.616481683
#>  [50,] 0.389182761 0.264580699
#>  [51,] 0.461186465 0.539316904
#>  [52,] 0.315241752 0.408232899
#>  [53,] 0.174675894 0.659546037
#>  [54,] 0.531573541 0.671235350
#>  [55,] 0.493637016 0.440323365
#>  [56,] 0.779308626 0.966028175
#>  [57,] 0.204178343 0.244085903
#>  [58,] 0.713397279 0.099581934
#>  [59,] 0.065216112 0.344393909
#>  [60,] 0.354206799 0.234373624
#>  [61,] 0.825199421 0.203312470
#>  [62,] 0.273818245 0.932182834
#>  [63,] 0.570044951 0.822999224
#>  [64,] 0.335719081 0.752364351
#>  [65,] 0.596262789 0.584152022
#>  [66,] 0.191518032 0.470415116
#>  [67,] 0.947763938 0.169543893
#>  [68,] 0.542480409 0.442107166
#>  [69,] 0.544603393 0.229656059
#>  [70,] 0.278597154 0.667078579
#>  [71,] 0.446702469 0.693002014
#>  [72,] 0.371511184 0.946442677
#>  [73,] 0.028060974 0.035634317
#>  [74,] 0.465987191 0.970663718
#>  [75,] 0.390031387 0.405425793
#>  [76,] 0.020065218 0.254152294
#>  [77,] 0.376970928 0.816199903
#>  [78,] 0.559912840 0.117174301
#>  [79,] 0.857083586 0.599873891
#>  [80,] 0.384809711 0.767122295
#>  [81,] 0.527917036 0.483729763
#>  [82,] 0.600637524 0.602794182
#>  [83,] 0.261371358 0.190460431
#>  [84,] 0.290050162 0.042055302
#>  [85,] 0.480075174 0.971886935
#>  [86,] 0.920005546 0.266935214
#>  [87,] 0.400720185 0.784096993
#>  [88,] 0.213172711 0.864106852
#>  [89,] 0.671766816 0.859260166
#>  [90,] 0.058614111 0.360923224
#>  [91,] 0.997069135 0.923350347
#>  [92,] 0.149035467 0.028260812
#>  [93,] 0.518556636 0.182138568
#>  [94,] 0.846120055 0.164809215
#>  [95,] 0.718269724 0.740783924
#>  [96,] 0.241314020 0.362488496
#>  [97,] 0.547043368 0.109307653
#>  [98,] 0.834801815 0.204481187
#>  [99,] 0.027956025 0.841935577
#> [100,] 0.469384300 0.261274201
#> [101,] 0.805680031 0.718780002
#> [102,] 0.814051310 0.290248573
#> [103,] 0.403911002 0.563401557
#> [104,] 0.218431009 0.797356727
#> [105,] 0.418361402 0.161560705
#> [106,] 0.668870748 0.337305192
#> [107,] 0.507650282 0.339792228
#> [108,] 0.660359306 0.039572344
#> [109,] 0.511791314 0.866419798
#> [110,] 0.835552437 0.497395365
#> [111,] 0.708781161 0.787274472
#> [112,] 0.874205941 0.765913061
#> [113,] 0.011479538 0.091098335
#> [114,] 0.888249568 0.979202033
#> [115,] 0.996346919 0.451903999
#> [116,] 0.500191501 0.220509015
#> [117,] 0.358967024 0.782125613
#> [118,] 0.774913022 0.071756529
#> [119,] 0.584475251 0.813438880
#> [120,] 0.633976371 0.235145431
#> [121,] 0.858666155 0.069741233
#> [122,] 0.566894339 0.809069437
#> [123,] 0.252997019 0.804349755
#> [124,] 0.918803215 0.971334850
#> [125,] 0.867350205 0.465993844
#> [126,] 0.248538697 0.160257567
#> [127,] 0.402881212 0.616006611
#> [128,] 0.769630176 0.836163560
#> [129,] 0.119485375 0.023908916
#> [130,] 0.194694962 0.677850582
#> [131,] 0.164569248 0.153888591
#> [132,] 0.663206581 0.042025983
#> [133,] 0.856575005 0.207839208
#> [134,] 0.926546448 0.738999184
#> [135,] 0.552377595 0.677837493
#> [136,] 0.577065694 0.389778267
#> [137,] 0.687447746 0.461410356
#> [138,] 0.244718230 0.672260357
#> [139,] 0.044617158 0.437008777
#> [140,] 0.909854557 0.938632902
#> [141,] 0.070681219 0.242959739
#> [142,] 0.996891473 0.662671473
#> [143,] 0.611852417 0.230870018
#> [144,] 0.172558846 0.477243469
#> [145,] 0.909440965 0.915830251
#> [146,] 0.037451167 0.282043500
#> [147,] 0.593553790 0.851486106
#> [148,] 0.236977555 0.546985629
#> [149,] 0.906297267 0.816824974
#> [150,] 0.818872984 0.481468417
#> [151,] 0.699829357 0.871804494
#> [152,] 0.220000329 0.699977908
#> [153,] 0.727990938 0.399034604
#> [154,] 0.217084462 0.098569190
#> [155,] 0.456230198 0.139947844
#> [156,] 0.332799758 0.794553529
#> [157,] 0.568352669 0.438554880
#> [158,] 0.252205725 0.057955800
#> [159,] 0.464013567 0.987365945
#> [160,] 0.917660507 0.668971011
#> [161,] 0.972844218 0.234047604
#> [162,] 0.819082447 0.608754554
#> [163,] 0.902923798 0.205866993
#> [164,] 0.581366044 0.967398173
#> [165,] 0.773008481 0.097154450
#> [166,] 0.995123026 0.248583801
#> [167,] 0.710971250 0.945975042
#> [168,] 0.214942596 0.965133131
#> [169,] 0.291757630 0.691834469
#> [170,] 0.721759729 0.902443844
#> [171,] 0.866615703 0.980848134
#> [172,] 0.238453106 0.039977679
#> [173,] 0.004496308 0.027085874
#> [174,] 0.943516464 0.703890758
#> [175,] 0.438137200 0.995084496
#> [176,] 0.750603328 0.397945047
#> [177,] 0.667815764 0.875892052
#> [178,] 0.407973201 0.556554881
#> [179,] 0.351248815 0.871745042
#> [180,] 0.738091561 0.583196370
#> [181,] 0.664285493 0.887495969
#> [182,] 0.085224700 0.765779263
#> [183,] 0.856132157 0.434412389
#> [184,] 0.076983322 0.000199395
#> [185,] 0.852844803 0.293527843
#> [186,] 0.106346961 0.916394638
#> [187,] 0.484802824 0.328257814
#> [188,] 0.247219110 0.865252975
#> [189,] 0.686569211 0.966725202
#> [190,] 0.163623198 0.005649085
#> [191,] 0.952824800 0.625096329
#> [192,] 0.321854551 0.225589029
#> [193,] 0.361534117 0.647449250
#> [194,] 0.887723417 0.773515690
#> [195,] 0.828014418 0.893856510
#> [196,] 0.100656458 0.225308117
#> [197,] 0.906051578 0.911397672
#> [198,] 0.772730364 0.868715968
#> [199,] 0.383370670 0.733856378
#> [200,] 0.999652457 0.087441138
dcMGL180.bivar(u1 = 0.001, u2 = 0.999, pars = 1)
#> [1] 0.05692685
pcMGL180.bivar(u1 = c(0.6, 0.1, 0.7), u2 = c(0.3, 0.6, 0.9), pars = 5)
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> For infinite domains Gauss integration is applied!
#> [1] 0.2791833 0.0928438 0.6965791
rcMGL180.bivar(n = 200, pars = 0.8)
#>               [,1]       [,2]
#>   [1,] 0.666176683 0.63818027
#>   [2,] 0.250727348 0.29563001
#>   [3,] 0.086841989 0.77289720
#>   [4,] 0.795670937 0.37627969
#>   [5,] 0.528244389 0.13119396
#>   [6,] 0.614816900 0.15905236
#>   [7,] 0.607248071 0.11639091
#>   [8,] 0.781661570 0.79972353
#>   [9,] 0.992791059 0.91190428
#>  [10,] 0.636287857 0.10241464
#>  [11,] 0.473893270 0.60264141
#>  [12,] 0.155836607 0.19948551
#>  [13,] 0.864008829 0.36330395
#>  [14,] 0.623532669 0.14644196
#>  [15,] 0.736691352 0.02687033
#>  [16,] 0.502147030 0.84090262
#>  [17,] 0.980474749 0.08690073
#>  [18,] 0.912816190 0.52763745
#>  [19,] 0.004037336 0.88503537
#>  [20,] 0.057867795 0.44549699
#>  [21,] 0.875626549 0.81658667
#>  [22,] 0.884810698 0.53203455
#>  [23,] 0.633311896 0.85516916
#>  [24,] 0.528959964 0.34576472
#>  [25,] 0.215827554 0.24282854
#>  [26,] 0.509607187 0.35859995
#>  [27,] 0.898049045 0.98020768
#>  [28,] 0.056540276 0.27788635
#>  [29,] 0.293860498 0.29788596
#>  [30,] 0.497735763 0.50569710
#>  [31,] 0.470065224 0.44014038
#>  [32,] 0.741573652 0.35536621
#>  [33,] 0.288246583 0.10132919
#>  [34,] 0.031537450 0.22256146
#>  [35,] 0.338922210 0.53106977
#>  [36,] 0.203749410 0.48968522
#>  [37,] 0.926790868 0.94333143
#>  [38,] 0.115257116 0.46673019
#>  [39,] 0.173299623 0.35581117
#>  [40,] 0.274236041 0.87156219
#>  [41,] 0.632670978 0.25689030
#>  [42,] 0.476571583 0.41126437
#>  [43,] 0.530121087 0.34117424
#>  [44,] 0.901433614 0.93272438
#>  [45,] 0.483965090 0.06356204
#>  [46,] 0.550266173 0.32917989
#>  [47,] 0.450538452 0.42601005
#>  [48,] 0.350493810 0.70147142
#>  [49,] 0.840965174 0.02210793
#>  [50,] 0.706249208 0.88299200
#>  [51,] 0.108715133 0.86614090
#>  [52,] 0.411921374 0.19677220
#>  [53,] 0.366768437 0.29886657
#>  [54,] 0.740490524 0.76204305
#>  [55,] 0.681787273 0.08533981
#>  [56,] 0.660746119 0.83491158
#>  [57,] 0.239238861 0.30525280
#>  [58,] 0.989127536 0.98447371
#>  [59,] 0.936862661 0.44403502
#>  [60,] 0.736158208 0.16857254
#>  [61,] 0.406163411 0.12909780
#>  [62,] 0.702439971 0.20901208
#>  [63,] 0.905377228 0.98627401
#>  [64,] 0.238744640 0.17265014
#>  [65,] 0.227523152 0.09531818
#>  [66,] 0.598009607 0.38777151
#>  [67,] 0.940917047 0.80075735
#>  [68,] 0.249752116 0.06723795
#>  [69,] 0.471677645 0.34906141
#>  [70,] 0.724194882 0.59122119
#>  [71,] 0.738762450 0.11422073
#>  [72,] 0.456168831 0.20804779
#>  [73,] 0.800999221 0.89475432
#>  [74,] 0.354945228 0.40458436
#>  [75,] 0.249760154 0.79701686
#>  [76,] 0.729022166 0.51995315
#>  [77,] 0.124378091 0.38455416
#>  [78,] 0.148256826 0.66826934
#>  [79,] 0.954150689 0.39043434
#>  [80,] 0.618943141 0.05155573
#>  [81,] 0.816687946 0.67313999
#>  [82,] 0.078200359 0.40888339
#>  [83,] 0.093813151 0.01523369
#>  [84,] 0.807056682 0.71456430
#>  [85,] 0.135624723 0.46320401
#>  [86,] 0.986727398 0.97885918
#>  [87,] 0.114339660 0.85004675
#>  [88,] 0.608457893 0.61109915
#>  [89,] 0.923823687 0.94932304
#>  [90,] 0.353218695 0.22598814
#>  [91,] 0.083346103 0.03640293
#>  [92,] 0.165491380 0.27794299
#>  [93,] 0.498752930 0.49069158
#>  [94,] 0.397855230 0.40997785
#>  [95,] 0.531040966 0.62738422
#>  [96,] 0.113625762 0.31065308
#>  [97,] 0.329451508 0.48006429
#>  [98,] 0.606222281 0.10974291
#>  [99,] 0.448709328 0.19371890
#> [100,] 0.757507324 0.87219008
#> [101,] 0.796791171 0.77123544
#> [102,] 0.931658297 0.48587803
#> [103,] 0.692726263 0.57109464
#> [104,] 0.006881906 0.26757191
#> [105,] 0.883690823 0.42105210
#> [106,] 0.284576923 0.87483231
#> [107,] 0.773891269 0.04849870
#> [108,] 0.866651452 0.84536507
#> [109,] 0.072718238 0.06078325
#> [110,] 0.105877343 0.42019104
#> [111,] 0.796335608 0.07804549
#> [112,] 0.743017081 0.59755087
#> [113,] 0.385994396 0.58253572
#> [114,] 0.558500812 0.49573825
#> [115,] 0.684405315 0.03547557
#> [116,] 0.898686239 0.83660769
#> [117,] 0.727353804 0.72812576
#> [118,] 0.346300603 0.12128808
#> [119,] 0.072129898 0.71682485
#> [120,] 0.973375330 0.45378635
#> [121,] 0.440508803 0.57661871
#> [122,] 0.145826348 0.65699814
#> [123,] 0.924116950 0.24484767
#> [124,] 0.794171144 0.89091534
#> [125,] 0.504110353 0.19158759
#> [126,] 0.738143056 0.67342766
#> [127,] 0.584038549 0.03441286
#> [128,] 0.560850000 0.84394588
#> [129,] 0.817622081 0.62566122
#> [130,] 0.921671734 0.68308020
#> [131,] 0.982230044 0.45238548
#> [132,] 0.271820027 0.92540920
#> [133,] 0.029154975 0.90467135
#> [134,] 0.464782412 0.18789753
#> [135,] 0.833509232 0.85426388
#> [136,] 0.941824349 0.97783688
#> [137,] 0.540860981 0.12729469
#> [138,] 0.377201179 0.86968546
#> [139,] 0.405180799 0.50947495
#> [140,] 0.256981763 0.02227061
#> [141,] 0.475719866 0.39107944
#> [142,] 0.503174750 0.22605644
#> [143,] 0.985281197 0.97329681
#> [144,] 0.042080875 0.31074232
#> [145,] 0.546910765 0.79010713
#> [146,] 0.018468151 0.47059097
#> [147,] 0.455294834 0.77841879
#> [148,] 0.639957599 0.85934992
#> [149,] 0.496853895 0.58517119
#> [150,] 0.925366368 0.09302620
#> [151,] 0.030208110 0.10264842
#> [152,] 0.351498686 0.66619021
#> [153,] 0.015434533 0.28488756
#> [154,] 0.463866743 0.82592894
#> [155,] 0.238940595 0.59716408
#> [156,] 0.435614273 0.33778488
#> [157,] 0.221545696 0.47578780
#> [158,] 0.360939011 0.25425626
#> [159,] 0.621039264 0.10897619
#> [160,] 0.640469329 0.87324753
#> [161,] 0.155768584 0.31042088
#> [162,] 0.543250819 0.14898297
#> [163,] 0.820965709 0.71538752
#> [164,] 0.033790804 0.31429802
#> [165,] 0.286588048 0.46835291
#> [166,] 0.975113969 0.97446357
#> [167,] 0.373543506 0.42910267
#> [168,] 0.172078678 0.77613276
#> [169,] 0.356930968 0.30510330
#> [170,] 0.291515738 0.77007228
#> [171,] 0.129272847 0.41600673
#> [172,] 0.718187418 0.90669196
#> [173,] 0.826775158 0.58852219
#> [174,] 0.890329981 0.97682320
#> [175,] 0.413777808 0.15570938
#> [176,] 0.149278673 0.62376612
#> [177,] 0.078473824 0.66076059
#> [178,] 0.324338098 0.53001452
#> [179,] 0.649411838 0.36993225
#> [180,] 0.991341925 0.20406956
#> [181,] 0.569101207 0.62786037
#> [182,] 0.275064903 0.31899342
#> [183,] 0.587743613 0.16926467
#> [184,] 0.449137605 0.79847171
#> [185,] 0.843639430 0.04366499
#> [186,] 0.956998707 0.97906121
#> [187,] 0.315355887 0.68191364
#> [188,] 0.485597503 0.33008575
#> [189,] 0.054312152 0.41849568
#> [190,] 0.156903058 0.65973024
#> [191,] 0.935116907 0.03857171
#> [192,] 0.237889085 0.29141645
#> [193,] 0.921615748 0.96305363
#> [194,] 0.866492337 0.48767260
#> [195,] 0.190303799 0.10291951
#> [196,] 0.440616761 0.06229158
#> [197,] 0.266270314 0.50442167
#> [198,] 0.650304524 0.50367328
#> [199,] 0.017331333 0.55577152
#> [200,] 0.591146120 0.72112409