Bivarite MGL-EV and survival MGL-EV copula

Density, distribution function, and h-function the bivariate MGL-EV and survival MGL-EV copula with copula parameter \(\delta\).

pcMGLEV.bivar(u1, u2, param)

dcMGLEV.bivar(u1, u2, param)

pcMGLEV180.bivar(u1, u2, param)

dcMGLEV180.bivar(u1, u2, param)

hcMGLEV180.bivar(u1, u2, param)

hcMGLEV.bivar(u1, u2, param)

Arguments

u1, u2
numeric vectors of equal length with values in \([0,1]\).

param
copula parameter, denoted by \(\delta>0\).

Value

dcMGLEV.bivar, pcMGLEV.bivar and hMGLEV.bivar give values of density distribution and h-function for the 2-dimensional MGL copula with copula parameter \(\delta>0\).

dcMGLEV180.bivar, pcMGLEV180.bivar and hMGLEV180.bivar give values of density and distribution and h-function for the 2-dimensional MGL copula with copula parameter \(\delta>0\).

hMGLEV.bivar and hMGLEV180.bivar return a list with

hfunc1: \(\partial C(u_1,u_2) / \partial u_1,\).

hfunc2: \(\partial C(u_1,u_2) / \partial u_2,\).

Details

The extreme value copula \(\bar{C}^{MGL-EV}\) of the survival MGL copula is given by $$ \bar{C}^{MGL-EV}(u_1,u_2;\delta)=\exp\left[\log\left(u_1u_2\right)A_{\delta}\left(\frac{\log\left(u_2\right)}{\log\left(u_1u_2\right)}\right)\right], $$ where \(A_{\delta}\) is the Pickands dependence function implemented by R's Afunction.

The cdf of the MGL-EV copula is also given by

$$ C^{MGL-EV}(u_1, u_2; {\delta})=\frac{(u_1+u_2)\left(1-A_{\delta}\left(\frac{u_1}{u_1+u_2}\right)\right)}{2\left(1 - A_{\delta}\left(\frac{1}{2}\right)\right)}. $$

The h-function is defined as the conditional distribution function of a bivariate copula, i.e., $$h_1(u_2|u_1,\delta) := P(U_2 \leq u_2 | U_1 = u_1) = \partial C(u_1,u_2) / \partial u_1,$$

$$h_2(u_1|u_2,\delta) := P(U_1 \leq u_1 | U_2 = u_2) := \partial C(u_1,u_2) / \partial u_2,$$

where \((U_1, U_2) \sim C\), and \(C\) is a bivariate copula distribution function with parameter(s) \(\delta\). For more details see Aas et al. (2009).

References

Aas, K., C. Czado, A. Frigessi, and H. Bakken (2009). Pair-copula constructions of multiple dependence. Insurance: Mathematics and Economics 44 (2), 182-198.

Examples

pcMGLEV.bivar(u1 = c(0.3, 0.9), u2 = c(0.5, 0.8), param = 2)

#> [1] 0.2586206 0.7602469

dcMGLEV.bivar(u1 = 0.001, u2 = 0.999, param = 1)

#> [1] 0.01644683

pcMGLEV180.bivar(u1 = c(0.3, 0.9), u2 = c(0.5, 0.8), param = 2)

#> [1] 0.2492633 0.7813228

dcMGLEV180.bivar(u1 = 0.5, u2 = 0.78, param = 1.8)

#> [1] 0.9513845

hcMGLEV180.bivar(u1 = 0.5, u2 = 0.78, param = 1.8)

#> $hfunc1
#> [1] 0.8908632
#> 
#> $hfunc2
#> [1] 0.2316631
#> 

hcMGLEV.bivar(u1 = 0.5, u2 = 0.78, param = 1.8)

#> $hfunc1
#> [1] 0.8314307
#> 
#> $hfunc2
#> [1] 0.2392104
#>